Refactor - move application subclass generation
The JerseyApplication dynamic subclassing should live alongside our other ASM usage in the proxy package.
